fix(qa): align Matrix migration with privatized helpers
