+++ extensions/qa-lab/src/live-transports/shared/live-artifacts.ts
@@ -1,13 +1,8 @@
 // Qa Lab plugin module implements live artifacts behavior.
-import { randomUUID } from "node:crypto";
 import { formatErrorMessage } from "openclaw/plugin-sdk/error-runtime";
 
 const REDACTED_QA_LIVE_LANE_ISSUE_DETAILS =
   "details redacted (OPENCLAW_QA_REDACT_PUBLIC_METADATA=1)";
-
-export function createLiveTransportQaRunId() {
-  return `${Date.now().toString(36)}-${randomUUID().slice(0, 8)}`;
-}
 
 export function redactQaLiveLaneDetails() {
   return REDACTED_QA_LIVE_LANE_ISSUE_DETAILS;
